Proto for presentation

+++ main.py
@@ -4,7 +4,8 @@ import time
 import gpio
 import datetime
 from utils import get_contours
-import glob_vars as gv
+import user_view as uv
+import notification as notify
 
 cv2.useOptimized()
 
@@ -24,36 +25,41 @@ def run():
     _, frame_old = cap.read()
     while True:
         _, frame = cap.read()
-
+        USER_SETTINGS = uv.get_vars()
         gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-        faces = [] if not gv.DETECT_FACES else face_cascade.detectMultiScale(gray, 1.1, 5)
-        bodies = [] if not gv.DETECT_BODIES else face_cascade.detectMultiScale(gray, 1.3, 5)
-        contours = [] if not gv.DETECT_CONTOURS else get_contours(frame, frame_old, gv.CONTOUR_MIN_AREA)
+        faces = [] if not USER_SETTINGS['DETECT_FACES'] else face_cascade.detectMultiScale(gray, 1.1, 5)
+        bodies = [] if not USER_SETTINGS['DETECT_BODIES'] else face_cascade.detectMultiScale(gray, 1.3, 5)
+        contours = [] if not USER_SETTINGS['DETECT_CONTOURS'] else get_contours(frame, frame_old, USER_SETTINGS['CONTOUR_MIN_AREA'])
 
+        print(len(faces), len(bodies), len(contours))
         if len(faces) + len(bodies) + len(contours) > 0:
             if detection:
                 timer_started = False
             else:
                 gpio.set_red_led(True)
                 detection = True
+                gpio.set_buzzer(True)
+                detection_stopped_time = time.time()
                 current_time = datetime.datetime.now().strftime("%d-%m-%Y-%H-%M-%S")
+                notify.email_alert(f"Wlamanie {current_time}", f"{current_time} rozpoczeto rejestrowac zdarzenie", "mail@temp")
+                notify.send_sms_via_email(f"{current_time} rozpoczeto rejestrowac zdarzenie", "mail@temp")
                 out = cv2.VideoWriter(f'{current_time}.mp4', fourcc, 20, frame_size)
         elif detection and out:
             if timer_started:
                 duration = time.time() - detection_stopped_time
-                if duration >= gv.SECONDS_RECORDING_AFTER_DETECTION:
+                if duration >= USER_SETTINGS['SECONDS_RECORDING_AFTER_DETECTION']:
                     detection = False
+                    gpio.set_buzzer(False)
                     timer_started = False
                     out.release()
                     out = None
                     gpio.set_red_led(False)
             else:
                 timer_started = True
-                detection_stopped_time = time.time()
         if detection:
             out.write(frame)
 
-        if len(sys.argv) == 1:
+        if len(sys.argv) == 2 and sys.argv[1] == 'debug':
             cv2.imshow('Camera view', frame)
             if cv2.waitKey(10) == ord('q'):
                 cv2.destroyAllWindows()
